refactor(tests): organize regression tests by domain, not severity
Redistribute tests from test_low_fixes.py and test_medium_fixes.py
into domain-appropriate locations:
- M1 (cancel_scan) -> core/resource_management/test_cancel_scan_warning.py
- M3 (engine_phase) -> core/test_engine_phase_events.py
- M6 (find_executable) -> utils/test_find_executable_fix.py
- M7 (glob recursive) -> utils/test_glob_recursive_fix.py
- L1 (mcp debug) -> cli/test_mcp_debug_prints_fix.py
- L2 (inspect split) -> cli/test_inspect_split_fix.py
- Dead code -> test_dead_code_removal.py

No behavior change. Same 1016 tests pass.
